+++ components/src/index.js
@@ -34,4 +34,8 @@ export { default as ListItem } from "./List/ListItem";
 export { default as List } from "./List/List";
 export { default as Textarea } from "./Textarea/Textarea";
 export { default as Tooltip } from "./Tooltip/Tooltip";
+export { default as Card } from "./Card/Card";
+export { default as CardSet } from "./Card/CardSet";
+export { default as CPCard } from "./Card/CPCard";
+
 export { default as DemoPage } from "./DemoPage/DemoPage";