+++ components/CHANGELOG.md
@@ -8,8 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
- - Custom link components supported in the NavBar.
+- Custom link components supported in the NavBar.
+- Added `Card` and `CardSet` components
+
 ### Changed
+- Label of `IconicButton` is now optional
 - Select API changed, given the following options:
 
     ```js